Replacing uint16 with uint32, adding encoder logic and changing answer from relative to absolute

+++ Main/src/odometry.h
@@ -4,7 +4,7 @@
 #include <Arduino.h>
 
 extern void     odometry_setup(void);
-extern uint16_t odometry_get_left_counter(void);
-extern uint16_t odometry_get_right_counter(void);
+extern uint32_t odometry_get_left_counter(void);
+extern uint32_t odometry_get_right_counter(void);
 
 #endif
